% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabuplot.R
\name{rabuplot}
\alias{rabuplot}
\title{Relative Abundance Plot}
\usage{
rabuplot(
  phylo_ob,
  predictor,
  type = "genus",
  xlabs = "Relative abundance (\%)",
  ylabs = "Average relative abundance",
  main = "Relative abundance plot",
  violin = TRUE,
  violin_scale = "width",
  legend_title = predictor,
  N_taxa = 15,
  No_other_type = FALSE,
  legend_names = NULL,
  Time = "Time",
  Timepoint = NULL,
  Strata = NULL,
  Strata_val = "1",
  No_legends = FALSE,
  No_names = FALSE,
  By_median = TRUE,
  Only_sig = FALSE,
  log = TRUE,
  log_max = 100,
  stat_out = FALSE,
  p_val = TRUE,
  colors = NULL,
  order = TRUE,
  reverse = FALSE,
  list_OTU = NULL,
  list_type = NULL,
  select_taxa = NULL,
  select_type = NULL,
  relative_abun = TRUE,
  bar_chart = FALSE,
  bar_wrap = NULL,
  bar_chart_stacked = TRUE,
  color_by = NULL,
  percent = FALSE,
  order_by = "Time",
  order_val = NULL,
  group = NULL,
  p_stars = FALSE,
  remove_collapsed_taxa = FALSE,
  stats = "non-parametric",
  p_adjust = FALSE
)
}
\arguments{
\item{phylo_ob}{Phyloseq object with metadata in sample_data}

\item{predictor}{Predictor for stastics/plotting}
}
\value{
A ggplot
}
\description{
For creating nice microbiome plots
}
